+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/keyvalue/api/cache/LockWatchValueCache.java
@@ -34,7 +34,7 @@ import java.util.Set;
 public interface LockWatchValueCache {
     void processStartTransactions(Set<Long> startTimestamps);
 
-    void updateCache(TransactionDigest digest, long startTs);
+    void updateCacheOnCommit(TransactionDigest digest, long startTs);
 
     TransactionScopedCache createTransactionScopedCache(long startTs);
 }
