+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/keyvalue/api/cache/TransactionCacheValueStore.java
@@ -1,0 +1,35 @@
+/*
+ * (c) Copyright 2021 Palantir Technologies Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.palantir.atlasdb.keyvalue.api.cache;
+
+import com.palantir.atlasdb.keyvalue.api.Cell;
+import com.palantir.atlasdb.keyvalue.api.CellReference;
+import com.palantir.atlasdb.keyvalue.api.TableReference;
+import java.util.Map;
+import java.util.Set;
+
+public interface TransactionCacheValueStore {
+    void cacheLocalWrite(TableReference tableReference, Cell cell, CacheValue value);
+
+    void updateLocalReads(TableReference tableReference, Map<Cell, byte[]> remoteReadValues);
+
+    void updateEmptyReads(TableReference tableReference, Set<CellReference> emptyCells);
+
+    Map<CellReference, CacheValue> getCachedValues(Set<CellReference> cellReferences);
+
+    Map<CellReference, CacheValue> getTransactionDigest();
+}
