+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/keyvalue/api/cache/TransactionScopedCacheImpl.java
@@ -21,34 +21,24 @@ import com.google.common.collect.Sets;
 import com.palantir.atlasdb.keyvalue.api.Cell;
 import com.palantir.atlasdb.keyvalue.api.CellReference;
 import com.palantir.atlasdb.keyvalue.api.TableReference;
-import com.palantir.atlasdb.keyvalue.api.cache.TransactionScopedCacheImpl.LocalCacheEntry.Status;
 import com.palantir.common.streams.KeyedStream;
-import java.util.HashMap;
 import java.util.Map;
-import java.util.Objects;
-import java.util.Optional;
 import java.util.Set;
 import java.util.function.BiFunction;
 import java.util.stream.Collectors;
-import org.immutables.value.Value;
 
 public final class TransactionScopedCacheImpl implements TransactionScopedCache {
-    private final ValueCacheSnapshot snapshot;
-    private final Map<CellReference, LocalCacheEntry> localUpdates;
+    private final TransactionCacheValueStore valueStore;
 
     public TransactionScopedCacheImpl(ValueCacheSnapshot snapshot) {
-        this.snapshot = snapshot;
-        localUpdates = new HashMap<>();
+        valueStore = new TransactionCacheValueStoreImpl(snapshot);
     }
 
     // TODO(jshah): figure out how we can improve perf given that this is now synchronised. Maybe its fine, but maybe
     //  we should instead autobatch?
     @Override
     public synchronized void write(TableReference tableReference, Cell cell, CacheValue value) {
-        CellReference cellReference = CellReference.of(tableReference, cell);
-        if (snapshot.isWatched(tableReference) && snapshot.isUnlocked(cellReference)) {
-            localUpdates.put(cellReference, LocalCacheEntry.write(value));
-        }
+        valueStore.cacheLocalWrite(tableReference, cell, value);
     }
 
     // TODO(jshah): as above. Equally, we should probably use the async value loader, then have it get afterwards,
@@ -61,74 +51,31 @@ public final class TransactionScopedCacheImpl implements TransactionScopedCache 
         Set<CellReference> cellReferences = cells.stream()
                 .map(cell -> CellReference.of(tableReference, cell))
                 .collect(Collectors.toSet());
-
-        // Read local values first - they may be different in the case of writes.
-        Map<CellReference, CacheValue> localCachedValues = getLocallyCachedValues(cellReferences);
-
-        // Filter out which values have not been read yet
-        Set<CellReference> thusFarUncachedValues = Sets.difference(cellReferences, localCachedValues.keySet());
-
-        // Read values from the snapshot. For the hits, mark as hit in the local map.
-        Map<CellReference, CacheValue> snapshotCachedValues = getSnapshotValues(thusFarUncachedValues);
-        snapshotCachedValues.forEach(
-                (cellReference, value) -> localUpdates.put(cellReference, LocalCacheEntry.hit(value)));
+        Map<CellReference, CacheValue> cachedValues = valueStore.getCachedValues(cellReferences);
 
         // Now we need to read the remaining values from the DB
-        Set<CellReference> nowActuallyUncachedValues =
-                Sets.difference(thusFarUncachedValues, localCachedValues.keySet());
+        Set<CellReference> uncachedCells = Sets.difference(cellReferences, cachedValues.keySet());
 
         Map<Cell, byte[]> remoteReadValues = valueLoader.apply(
-                tableReference,
-                nowActuallyUncachedValues.stream().map(CellReference::cell).collect(Collectors.toSet()));
+                tableReference, uncachedCells.stream().map(CellReference::cell).collect(Collectors.toSet()));
 
-        KeyedStream.stream(remoteReadValues)
-                .mapKeys(cell -> CellReference.of(tableReference, cell))
-                .map(CacheValue::of)
-                .filterKeys(snapshot::isUnlocked)
-                .forEach((cell, value) -> localUpdates.put(cell, LocalCacheEntry.read(value)));
+        valueStore.updateLocalReads(tableReference, remoteReadValues);
 
         // The get method does not return an entry if a value is absent; we want to cache this fact.
-        Set<CellReference> emptyCells = Sets.difference(nowActuallyUncachedValues, remoteReadValues.keySet());
-        emptyCells.stream()
-                .filter(snapshot::isUnlocked)
-                .forEach(cell -> localUpdates.put(cell, LocalCacheEntry.read(CacheValue.empty())));
+        Set<CellReference> emptyCells = Sets.difference(uncachedCells, remoteReadValues.keySet());
+
+        valueStore.updateEmptyReads(tableReference, emptyCells);
 
         // Now we need to combine all the maps, filtering out empty values
-        ImmutableMap.Builder<Cell, byte[]> output = ImmutableMap.builder();
-        output.putAll(remoteReadValues);
-        output.putAll(filterEmptyValues(snapshotCachedValues));
-        output.putAll(filterEmptyValues(localCachedValues));
-
-        return output.build();
+        return ImmutableMap.<Cell, byte[]>builder()
+                .putAll(remoteReadValues)
+                .putAll(filterEmptyValues(cachedValues))
+                .build();
     }
 
     @Override
     public synchronized TransactionDigest getDigest() {
-        // The digest doesn't contain values cached locally due to writes because they will be filtered out based on
-        // locks anyway.
-        Map<CellReference, CacheValue> loadedValues = KeyedStream.stream(localUpdates)
-                .filter(entry -> entry.status().equals(Status.READ))
-                .map(LocalCacheEntry::value)
-                .collectToMap();
-        return TransactionDigest.of(loadedValues);
-    }
-
-    private Map<CellReference, CacheValue> getSnapshotValues(Set<CellReference> thusFarUncachedValues) {
-        return KeyedStream.of(thusFarUncachedValues)
-                .map(snapshot::getValue)
-                .filter(Optional::isPresent)
-                .map(Optional::get)
-                .filter(CacheEntry::isUnlocked)
-                .map(CacheEntry::value)
-                .collectToMap();
-    }
-
-    private Map<CellReference, CacheValue> getLocallyCachedValues(Set<CellReference> cellReferences) {
-        return KeyedStream.of(cellReferences)
-                .map(localUpdates::get)
-                .filter(Objects::nonNull)
-                .map(LocalCacheEntry::value)
-                .collectToMap();
+        return TransactionDigest.of(valueStore.getTransactionDigest());
     }
 
     private Map<Cell, byte[]> filterEmptyValues(Map<CellReference, CacheValue> snapshotCachedValues) {
@@ -137,39 +84,5 @@ public final class TransactionScopedCacheImpl implements TransactionScopedCache 
                 .map(value -> value.value().get())
                 .mapKeys(CellReference::cell)
                 .collectToMap();
-    }
-
-    @Value.Immutable
-    public interface LocalCacheEntry {
-        Status status();
-
-        CacheValue value();
-
-        static LocalCacheEntry read(CacheValue value) {
-            return ImmutableLocalCacheEntry.builder()
-                    .status(Status.READ)
-                    .value(value)
-                    .build();
-        }
-
-        static LocalCacheEntry write(CacheValue value) {
-            return ImmutableLocalCacheEntry.builder()
-                    .status(Status.WRITE)
-                    .value(value)
-                    .build();
-        }
-
-        static LocalCacheEntry hit(CacheValue value) {
-            return ImmutableLocalCacheEntry.builder()
-                    .status(Status.HIT)
-                    .value(value)
-                    .build();
-        }
-
-        enum Status {
-            READ,
-            WRITE,
-            HIT;
-        }
     }
 }
